# src/docs/openapi-docs.json
{"openapi":"3.0.0","info":{"version":"1.0.0","title":"Todo List Api","description":"Todo List Api Description"},"servers":[{"url":"http://localhost:5000"}],"components":{"schemas":{"User":{"type":"object","properties":{"email":{"type":"string","format":"email"},"username":{"type":"string"},"password":{"type":"string"}},"required":["email","username","password"]},"UserLoginInput":{"type":"object","properties":{"username":{"type":"string","minLength":1,"maxLength":100},"password":{"type":"string","minLength":1,"maxLength":100}},"required":["username","password"]},"UserSignUpInput":{"allOf":[{"$ref":"#/components/schemas/User"},{"type":"object","properties":{"confirm_password":{"type":"string"}},"required":["confirm_password"]}]},"Session":{"type":"object","properties":{"data":{"type":"object","properties":{"email":{"type":"string","format":"email"},"username":{"type":"string"},"id":{"type":"number"}},"required":["email","username","id"]}},"required":["data"]},"SuccessReponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"}},"required":["success","message"]},"ValidationError":{"type":"object","properties":{"message":{"type":"string"},"statusCode":{"type":"number"},"validationErrors":{"type":"array","items":{"type":"object","properties":{"code":{"type":"string"},"validation":{"type":"string"},"message":{"type":"string"},"path":{"type":"array","items":{"type":"string"}}},"required":["code","validation","message","path"]}}},"required":["message","statusCode","validationErrors"]},"NotFoundError":{"type":"object","properties":{"message":{"type":"string","example":"Not Found"},"statusCode":{"type":"number","example":404}},"required":["message","statusCode"]},"ConflictError":{"type":"object","properties":{"message":{"type":"string","example":"Conflict"},"statusCode":{"type":"number","example":409}},"required":["message","statusCode"]},"UnauthorizedError":{"type":"object","properties":{"message":{"type":"string","example":"Unauthorized"},"statusCode":{"type":"number","example":401}},"required":["message","statusCode"]},"ServerInternalError":{"type":"object","properties":{"message":{"type":"string","example":"Server Internal Error"},"statusCode":{"type":"number","example":500}},"required":["message","statusCode"]},"CreatePostInput":{"type":"object","properties":{"content":{"type":"string","minLength":1,"maxLength":5000},"audience":{"type":"string","enum":["public","private"],"default":"public"}},"required":["content"]},"Post":{"allOf":[{"$ref":"#/components/schemas/CreatePostInput"},{"type":"object","properties":{"id":{"type":"number"},"user_id":{"type":"number"},"likes":{"type":"number"},"comments":{"type":"number"},"shares":{"type":"number"}},"required":["id","user_id","likes","comments","shares"]}]},"PostComment":{"type":"object","properties":{"comment":{"type":"string","minLength":1}},"required":["comment"]}},"parameters":{}},"paths":{"/api/v1/auth/login":{"post":{"tags":["Authentication"],"operationId":"postLoginUser","summary":"Login User","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginInput"}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessReponse"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConflictError"}}}},"500":{"description":"Server Internal Error.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerInternalError"}}}}}}},"/api/v1/auth/signup":{"post":{"tags":["Authentication"],"operationId":"postSignUpUser","summary":"Sign Up User","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignUpInput"}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessReponse"}}}},"400":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationError"}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConflictError"}}}},"500":{"description":"Server Internal Error.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerInternalError"}}}}}}},"/api/v1/auth/session":{"get":{"tags":["Authentication"],"operationId":"getCurrentSession","summary":"Get User Session","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Session"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"500":{"description":"Server Internal Error.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerInternalError"}}}}}},"delete":{"tags":["Authentication"],"operationId":"deleteCurrentSession","summary":"Delete Current Session","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Session"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"500":{"description":"Server Internal Error.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerInternalError"}}}}}}},"/api/v1/post":{"post":{"tags":["Post"],"operationId":"postCreatePost","summary":"Create Post","requestBody":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePostInput"}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessReponse"}}}},"400":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationError"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"500":{"description":"Server Internal Error.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerInternalError"}}}}}}},"/api/v1/post/{postId}":{"get":{"tags":["Post"],"operationId":"getUserPost","summary":"Get User Post","parameters":[{"schema":{"type":"number"},"required":true,"name":"postId","in":"path"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/Post"}},"required":["data"]}}}},"401":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"500":{"description":"Server Internal Error.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerInternalError"}}}}}}},"/api/v1/post/all/{userId}":{"get":{"tags":["Post"],"operationId":"getUserPosts","summary":"Get User Posts","parameters":[{"schema":{"type":"number"},"required":true,"name":"userId","in":"path"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/Post"}}},"required":["data"]}}}},"401":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"500":{"description":"Server Internal Error.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerInternalError"}}}}}}},"/api/v1/post/like/{postId}":{"post":{"tags":["Post"],"operationId":"postLikePost","summary":"Like User Posts","parameters":[{"schema":{"type":"number"},"required":true,"name":"postId","in":"path"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessReponse"}}}},"401":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"500":{"description":"Server Internal Error.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerInternalError"}}}}}}},"/api/v1/post/comment/{postId}":{"post":{"tags":["Post"],"operationId":"postPostComment","summary":"Comment To User Posts","parameters":[{"schema":{"type":"number"},"required":true,"name":"postId","in":"path"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessReponse"}}}},"401":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedError"}}}},"500":{"description":"Server Internal Error.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerInternalError"}}}}}}}}}